Отделил зерна от плевел (Вынес устаревшие методы, чтобы не мешались)
