+++ VkNet/Categories/StatusCategory.cs
@@ -1,18 +1,15 @@
-﻿using System.Collections.Generic;
-using VkNet.Enums.Filters;
-using VkNet.Model.Attachments;
+﻿using VkNet.Enums.Filters;
 
 namespace VkNet.Categories
 {
-	using System;
-	using JetBrains.Annotations;
+    using JetBrains.Annotations;
 	using Model;
 	using Utils;
 
 	/// <summary>
 	/// Методы для работы со статусом пользователя или сообщества.
 	/// </summary>
-	public class StatusCategory
+	public partial class StatusCategory
 	{
 		private readonly VkApi _vk;
 
@@ -68,28 +65,6 @@ namespace VkNet.Categories
 			};
 
 			return _vk.Call("status.set", parameters);
-		}
-
-		/// <summary>
-		/// Устанавливает новый аудиозапись текущему пользователю. 
-		/// </summary>
-		/// <param name="audio">
-		/// Текущая аудиозапись, которую необходимо транслировать в статус, задается в формате oid_aid (
-		/// идентификатор владельца и идентификатор аудиозаписи, разделенные знаком подчеркивания). 
-		/// Для успешной трансляции необходимо, чтобы она была включена пользователем, в противном случае будет возвращена 
-		/// ошибка 221 ("User disabled track name broadcast"). При указании параметра audio параметр text игнорируется.
-		/// </param>
-		/// <returns>Возвращает true, если статус был успешно установлен, false в противном случае.</returns>
-		/// <remarks>
-		/// Для вызова этого метода Ваше приложение должно иметь права с битовой маской, содержащей <see cref="Settings.Status"/>. 
-		/// Страница документации ВКонтакте <see href="http://vk.com/dev/status.set"/>.
-		/// </remarks>
-		[ApiVersion("5.44")]
-		[Obsolete("Данный метод устарел. Пожалуйста используйте метод Audio.SetBroadcast")]
-		public bool Set([NotNull] Audio audio)
-		{
-			_vk.Audio.SetBroadcast(string.Format("{0}_{1}", audio.OwnerId, audio.Id), new List<long>());
-			return true;
 		}
 	}
 }
